requirements.txt
openai-whisper>=20231117
openai>=1.0.0
streamlit>=1.0.0
python-dotenv>=0.19.0
PyYAML>=6.0
pathlib>=1.0.1
pandas>=2.0.0
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1 